modules/hunter.py

import requests
import socket

try:
    from googlesearch import search
except ImportError:
    search = None


class SunkarHunter:
    def __init__(self):
        # Real OSINT Footprints (Dorks) targeting known scam patterns in KZ
        self.dorks = [
            'site:.xyz "kaspi" "бонус"',
            'inurl:aviator "казахстан" "ойнау"',
            '"финансовая пирамида" сайт инвест казахстан',
            'site:.online "halyk" "выплата"',
            '"invest" "kz" "пассивный доход" site:.com',
            '"aviator" "получить" "kz" site:.top',
        ]

    def domain_resolves(self, url):
        """DNS check — domain must actually exist."""
        try:
            domain = url.replace("https://", "").replace("http://", "").split("/")[0]
            socket.gethostbyname(domain)
            return True
        except socket.gaierror:
            return False

    def check_live(self, url):
        """HTTP check — server must respond."""
        try:
            r = requests.get(
                url, timeout=3, headers={"User-Agent": "Mozilla/5.0"}, allow_redirects=True
            )
            return r.status_code < 400
        except Exception:
            return False

    def proactive_search(self):
        """
        Real-time discovery via Google Dorking.
        Every returned URL has passed DNS + HTTP validation.
        """
        discovered = []

        if not search:
            return []

        for dork in self.dorks:
            try:
                for url in search(dork, num_results=5, lang="kk"):
                    if url not in discovered and self.domain_resolves(url):
                        discovered.append(url)
            except Exception:
                # Google may rate-limit — skip and continue
                continue

        return discovered
